fixed href for stylesheet
I removed the "/" in front of style.css from the link to your stylesheet. It was not allowing the CSS stylesheet to be linked, and work on https://archsofia.github.io/ToDoList_JS/ . Now you should be able to see the styles applied on your live site for this app.